+++ Source/NSObject.m
@@ -108,7 +108,7 @@ static Class	NSConstantStringClass;
 @class	NSDataMalloc;
 @class	NSMutableDataMalloc;
 
-GS_ROOT_CLASS @interface	NSZombie
+GS_ROOT_CLASS @interface NSZombie
 {
   Class	isa;
 }
@@ -1837,7 +1837,7 @@ static id gs_weak_load(id obj)
         [NSException
 	  raise: NSGenericException
 	  format: @"Autorelease would release object too many times.\n"
-	  @"%d release(s) versus %d retain(s)", release_count, retain_count];
+	  @"%lu release(s) versus %lu retain(s)", (unsigned long)release_count, (unsigned long)retain_count];
     }
 
   (*autorelease_imp)(autorelease_class, autorelease_sel, self);
